+++ CHANGELOG.md
@@ -153,3 +153,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - DataSHIELD Project Status Error
 - Catch token manager exceptions and return ERROR status
 - Send emails in async execution pool
+- Notification async execution pool